LU-6205 tests: fix bash expansion of fid
When calling lfs path2fid, the FID is returned between bracket. When
that fid variable is used, it is then expanded by the shell to
something else. For instance when it is given to fid2path, it returns
an error:

  $ ../utils/lfs fid2path lustre [0x200000be7:0xb:0x0]
  bad FID format [x], should be [0x1:0x2:0x0]

  fid2path: error on FID x: Invalid argument

This will cause some tests, like 154A or 238, to sometimes fail.

Trim the brackets to fix the problem, like it is done in some other
tests.

And replace "$(lfs ..." with "$($LFS ...".

Signed-off-by: frank zago <fzago@cray.com>
Change-Id: I3c1a34585ebaa596d66063f5ada3ccfc4d202ade