LU-4582 mgc: replace hard-coded MGC_ENQUEUE_LIMIT value
During client mount, the client will send an LDLM_ENQUEUE request to
MGS with send delay set to MGC_ENQUEUE_LIMIT, which is hard coded to
50 seconds. On the other hand, the interval for pinger is deduced from
obd_timeout. When obd_timeout is configured for a longer period of
time, so does the pinger. We know that connecting to the secondary MGS
node is triggered by the pinger. Now that we have a longer interval,
the pinger will not be able to try the secondary before the
LDLM_ENQUEUE request fails the mount using the same delay limit.

This code change will replace the hard-coded send delay being
mentioned above with a value that is long enough to give the client a
chance to connect to the secondary MGS if exists.

This patch is back-ported from the following one:
Lustre-commit: fb7ac6a880b71cb9e0252bcbb6d9bbb4b4ddf7fd
Lustre-change: http://review.whamcloud.com/9217

Xyratex-bug-id: MRP-1516

Signed-off-by: Cheng Shao <cheng_shao@xyratex.com>
Change-Id: I3490552b54d7978fe63478951bb166b77254dc3d