LU-6234 util: check fsname and pool name
The pool name length check should not include its fsname. This patch
get rid of the possibly added fsname in the pool name argument.

Check lustre filesystem name (fsname) in mkfs.lustre, it can only
contains alphabetic, digital char and '_'/'-'.

Check pool name in lctl/lfs, its character also abides by the
constraint of the fsname.

Add test cases to check the pool names for length, with and without
a filesystem name prefix, and with invalid names.

Also fix pool_add in the test-framework so that it greps for the
poolname plus the filesystem name, so it doesn't match more lines than
necessary.

Signed-off-by: Bobi Jam <bobijam.xu@intel.com>
Signed-off-by: frank zago <fzago@cray.com>
Change-Id: I4f542e1bc7b3ea97b35ec60e51aa971e73da91eb