LU-5319 ptlrpc: atomically assign XID and put on the sending list
it's required to have all assigned XID's on the sending list so
we can learn the lowest active XID at any point. this is in turn
required to be able to report the highest replied XID to the server
so the server can release corresponding slots in last_rcvd.

Test-Parameters: fortestonly mdtfilesystemtype=ldiskfs mdsfilesystemtype=ldiskfs \
ostfilesystemtype=ldiskfs envdefinitions=SLOW=yes

Change-Id: I1ece2f1806386a657a9a57d4fc3c8ffb2f32b47a
Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>