LU-6319 tests: Check if ALWAYS_EXCEPT tests pass
Many tests listed in the ALWAYS_EXCEPT list were added, in some
cases, years ago. These tests need to be retested with the current
version of Lustre to see which tests are able to run again and
which need to remain on the ALWAYS_EXCEPT list.

Test-Parameters: fortestonly testlist=sanityn,sanity,recovery-small,replay-single
Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Ibf764f6cfc50ca03dda9602011faf30864715a8b