LU-6319 tests: Check if ALWAYS_EXCEPT tests pass
In many of the Lustre test suites, tests listed in the ALWAYS_EXCEPT
list were added, in some cases, years ago. These tests were retested
with the current version of Lustre, pass testing and are being removed
from the ALWAYS_EXCEPT list.

All tests listed in the ALWAYS_EXCEPT list for recovery-small,
replay-single, sanity and some in sanityn are re-enabled.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Ibf764f6cfc50ca03dda9602011faf30864715a8b