LU-6341 llite: not check security for default LMVEA
Skip the security check for default stripeEA, if it is
from ll_new_node.

When creating a sub-directory under the directory with
default LMV stripeEA, client needs to cache the default
stripeEA on the client side, otherwise client might
send the create request to the wrong MDT. So
ll_getxattr_common() should not check security in this
case, otherwise only admin will be able to create new
dir under such directory.

Signed-off-by: wang di <di.wang@intel.com>
Change-Id: I0d8884b78fbc8b8d1930b1133150686e65d20c54