LU-6325 ptlrpc: CPT-aware placement of ptlrpc_request_set
Ensure that the ptlrpc_request_set for a ptlrpcd thread is created on
the same CPT that the thread will work on. When threads are bound to
specific nodes and/or CPUs in a NUMA system, it pays to ensure that
the datastructures used by these threads are also on the same node.

Implementation notes:

* ptlrpc_prep_set_cpt(): New function, allocates a ptlrpc_request_set
  on the specified CPT.

* ptlrpc_prep_set(): Changed to forward to ptlrpc_prep_set_cpt(). It
  specifies the use of the current CPT.

This change builds on
  I3ac40ea56f9c792c3e7c36967e2e1f20105c566c

Tested-by: Stephen Champion <schamp@sgi.com>
Signed-off-by: Olaf Weber <olaf@sgi.com>
Change-Id: I38c5e40cdefc8c22ca697239004342c4c36a58ad