LU-0000 ptlrpc: fix NUMA binding policy to use CPT
The ptlrpcd binding policy PDB_POLICY_NEIGHBOR does not correctly
manage offline cpus when kernel supports NUMA.

This patch implements a separate policy PDB_POLICY_CPT that relies
on the libcfs compute partition table. With that policy, each ptlrpcd
thread is bound to a cpu of the cptable. Its partners are all the
threads bound to a cpu of the same cpt. Additionnaly this policy take
care of NUMA binding.

Signed-off-by: Gregoire Pichon <gregoire.pichon@bull.net>
Change-Id: Ib3b3f15c1b19beeda3a7fa40a438a6dc07031974