LU-6254 ldiskfs: only remove page from cache after I/O
When read cache or write through cache is not enabled for OSD,
pages might be removed from cache when preparing the I/O. This
isn't a good behavior especially when using features like
ladvise.

This patch change the behavior so that pages will be removed
from cache when being released. This patch also cleanups the
proc entries of read cache and write through cache.

Signed-off-by: Li Xi <lixi@ddn.com>
Change-Id: I1d6a3e43d8588ca345fef547daaea070f42a19c4