LU-6388 utils: make llog_reader parse changelogs
Improve support for changelog_catalog records and changelog records.

Extend llog_reader to handle records of type CHANGELOG_REC or
CHANGELOG_USER_REC (present in lustre logs with changelog entries).

Records in the changelog_catalog (type LLOG_LOGID_MAGIC) contain the
ID of the object (file) containing actual changelog records.  The new
output provides the pathname to those files, as one would see them
when mounting an ldiskfs or zfs target via the posix layer. This makes
it easy to find those files to inspect them.

Move some logid/fid manipulation functions, which are specific to
llogs, out of lustre_log.h into a new header lustre_log_user.h.  One
of the functions is necessary for llog_reader to display useful
information for certain lustre log record types.  They are all related
to logid/fid conversion or access, and are not dependent on dt_object
like the rest of the material in lustre_log.h.

Test the size of the file being read.  If it is too small to contain
an llog_log_hdr, exit.  It is either corrupt, in which case parsing
will fail, or is not an llog file.

Change-Id: I7120c391702fd5b8e18031bbe6e109d9e079690f
Signed-off-by: Olaf Faaland <faaland1@llnl.gov>