LU-6325 libcfs: easy config for NUMA-match CPT table
If user wants to create CPT table that can match numa topology,
she has to query cpu & numa topology, then provide a pattern
string to describe the topology, this is inconvenient.

To improve it, this patch can create NUMA-match CPT table if
cpu_pattern is set to "N" or "n".

Signed-off-by: Liang Zhen <liang.zhen@intel.com>
Change-Id: I608f47ad6856ded5bf2f5f223b77b02906ebc8cc