LU-6325 ptlrpc: make ptlrpcd threads cpt-aware
On NUMA systems, the placement of worker threads relative to the
memory they use greatly affects performance. The CPT mechanism can be
used to constrain a number of Lustre thread types, and this change
makes it possible to confure the placement of ptlrpcd threads in a
similar manner.

To simplify the code changes, the global structures used to manage
ptlrpcd threads are changed to one per CPT. In particular this means
there will be one ptlrpcd recovery thread per CPT.

To prevent ptlrpcd threads from wandering all over the system, all
ptlrpcd thread are bound to a CPT. Note that some CPT configuration
is always created, but the defaults are not likely to be correct for
a NUMA system.

Ensure that the ptlrpc_request_set for a ptlrpcd thread is created on
the same CPT that the thread will work on. When threads are bound to
specific nodes and/or CPUs in a NUMA system, it pays to ensure that
the datastructures used by these threads are also on the same node.

Visible changes:

* ptlrpcd thread names include the CPT number, examples are
  "ptlrpcd_02_rcv" for a recovery thread, and "ptlrpcd_02_07"
  for a worker thread. In this case the "07" is relative to
  the CPT, and not a CPU number.

Tunables changed:

* max_ptlrpcds: This now controls the maximum number of ptlrpcd
  threads per CPT, as opposed to a global limit.

Tunables added:

* ptlrpcd_cpts (string): A CPT string describing the CPU partitions
  that ptlrpcd threads should run on. Used to make ptlrpcd threads
  run on a subset of all CPTs.

* num_ptlrpcd_partners (int): The desired number of partner threads
  for each ptlrpcd thread. Default is 1, corresponding to the old
  PDB_POLICY_PAIR. A negative value makes all ptlrpcd threads in a
  CPT partners of each other.

Tunables removed:

* ptlrpcd_bind_policy: The new num_ptlrpcd_partners can be used to
  obtain the same settings.

Internal interface changes:

* pdb_policy_t and related code have been removed. Groups of partner
  ptlrpcd threads are still created, and all threads in a partner
  group are bound on the same CPT. The ptlrpcd threads bound to a
  CPT are typically divided into several partner groups. The partner
  groups on a CPT all have an equal number of ptlrpcd threads.

* pdl_policy_t and related code have been removed. Since ptlrpcd
  threads are not bound to a specific CPU, all the code that avoids
  scheduling on the current CPU (or attempts to do so) has been
  removed as non-functional. A simplified form of PDL_POLICY_LOCAL
  is kept as the only load policy.

* LIOD_BIND and related code have been removed. All ptlrpcd threads
  are now bound to a CPT, and no additional binding policy is
  implemented.

* ptlrpc_prep_set_cpt(): New function, allocates a ptlrpc_request_set
  on the specified CPT.

* ptlrpc_prep_set(): Changed to forward to ptlrpc_prep_set_cpt(). It
  specifies the use of the current CPT.

Signed-off-by: Olaf Weber <olaf@sgi.com>
Change-Id: I3ac40ea56f9c792c3e7c36967e2e1f20105c566c