LU-6383 ldlm: Improve debug for LDLM lock matching
The search_queue function is at the core of ldlm_lock
matching, and there is currently no debug output from
it to track the lock matching process and indicate why
lock matching failed.

This is extremely useful in debugging certain issues,
such as LU-6368.

This patch adds debug in search_queue.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I8a7af6f119ab5425c2f4275e82cfd2241399857b