LU-5614 build: add kmod for client
When the Lustre is built as client with driver update support,
kmod package has ability to check and copy modules for the new
added kernel. The package includes dependencies on a kernel symbols
version. In most cases this is enough for a minor security kernel
updates. New param build_kmod was added to spec file. For Lustre
client build_kmod is enabled by default, and disabled for servers.

This patch changes the Lustre modules location to
[extra|updates]/lustre-client/[fs|net]/ for client build.
Kmod utils/scripts requires this modification, cause they use %name
for postinstall scripts. Path to Lustre modules for server build
stay the same.

Signed-off-by: Alexander Boyko <alexander.boyko@seagate.com>
Xyratex-bug-id: MRP-1759
Change-Id: I159c2d29577a605c0c96e0c9d93a86e7518b28fb