LU-6409 nodemap: replace mutex with spinlock
In order to avoid deadlock, calls to nm_member_reclassify_nodemap
need to be serialized. Previously this was was done by a mutex,
however this is not safe to do in a hash callback. This patch
converts the mutex to a spinlock. The lock is only taken during
configuration changes, which are rare, and so it shouldn't impact
performance.

Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: I1d16f61ba7739af0494944f6515508b3f0d19ba1