LU-6418 llite: Simplify code of ll_direct_rw_pages
In function ll_direct_rw_pages, the cl_page_find will return a
cl_page with cp_type with value of CPT_TRANSIENT for sure. And
it will not cache this type of cl_page anymore. This patch remove
the code of logical when cp_type is equal to CPT_CACHEABLE which
is needed in old version.

Signed-off-by: Wu Libin <lwu@ddn.com>
Change-Id: I6634c31dca65a54ad6d0e092554c033403315077