LU-6485 libcfs: distinguish kernelcomm by serial
currently, different lustre_kernelcomm is distinguished by
"uid" and "group", which is insufficient. for instance,
the "group" for HSM case is always be KUC_GRP_HSM(0x02),
and the lustre_kernelcomm(s) will be regarded to be the same
if they are created by the same user.

this patch introduces a serial for lustre_kernelcomm to make
it to be different from others.

Change-Id: Iaf5e9fbb9bad2540f51da2c4fd9c4047640d0877
Signed-off-by: Hongchao Zhang <hongchao.zhang@intel.com>