LU-6475 mdt: race between open and migrate
During intent open, if it founds the parent has been
migrated to another MDT, it should retry the open
request with the new object, so it needs to keep the
old object in the orphan list, which will be cleanup
during next recovery. Note: if the client still using
the old FID after next recovery, it will return -ENOENT
for the application.

Add sanityn.sh 80b to test race between open/stat and
migration.

Because migration will change the FID of the object,
so CLIO needs to rebuild the top/sub object cache anyway
But in current implemenation, client will build the
top/sub object cache even without layout lock, and if
client does not have layout lock, then the MDT can not
notify client to invalid its layout. So this patch will
also let CLIO to rebuild the object cache whenever it
find the layout is inconsistent, instead of only with
lo_layout_invalid flag.

Signed-off-by: wang di <di.wang@intel.com>
Change-Id: Iad3e5006079e9f72b338680e17448c39c6d1af98