LU-6495 o2iblnd: IB Multi-Rail
Code contributed by Fujitsu, documentation below:

Steps to set IB Multi-Rail

0. Setting IB
============

Please configure all IB devices in the unique subnet.
Please allocate IP address of IPoIB to each IB HCA device.

1. Preparing /etc/sysconfig/lnet_o2ibs.conf configuration file
====================================================

Please create the file /etc/sysconfig/lnet_o2ibs.conf in the following format.

192.168.128.10@o2ib0 192.168.128.10 192.168.128.11
192.168.128.12@o2ib0 192.168.128.12 192.168.128.13
192.168.128.14@o2ib0 192.168.128.14 192.168.128.15

Please list all IP addresses of a NID on a single line.
Set the first column to NID, and then set the second and the following columns to IP addresses.
Please describe information about all nodes which have multiple IB HCAs in this file.

Please describe IP addresses in the order in which the devices are specified in the networks parameter of lnet module.
For example, when networks parameter is networks=o2ib0(ib0,ib2), please specify IP addresses in this file in the order of ib0, ib2.
Moreover, please specify the NID based on the IP address in the second column as the NID in the first column.
For example, when 192.168.128.10 is specified in the second column, the first column should be set to 192.168.128.10@o2ib0

All nodes which have IB HCA in the system need to have this configuration file with unique contents.

2. Starting lnet
==============
Please start lnet by executing "service lnet start".
The setting of IB Multi-Rail will be done in the service script of lnet.

Change-Id: I343a0f61d213dadcae73bdbe10b285a896449529
Sifned-off-by: Oleg Drokin <oleg.drokin@intel.com>