LU-6302 lov: copy_to_user uses wrong casting
While testing on newer kernels lov_obd.c failes to compile
with the following warning.

In function ‘copy_to_user’,
    inlined from ‘lov_iocontrol’ at
    lustre-2.7.52/lustre/lov/lov_obd.c:1168:
./arch/x86/include/asm/uaccess.h:735: error: call to
‘__copy_to_user_overflow’ declared with attribute warning:
copy_to_user() buffer size is not provably correct

In lov_iocontrol the data was being casted to int instead
of the required unsigned long. This patch changes the cast
to what is needed for copy_to_user.

Change-Id: I75b9f4f935d57dc3df56a2c0869c7d2506217d39
Signed-off-by: James Simmons <uja.ornl@yahoo.com>