LU-6531 lnet: o2iblnd Channel Bonding Solution
Land Fujitsu's Channel Bonding solution to lustre
community tree.

Fujitsu implemented Channel bonding entirely in the iblnd layer.

An overview of their solution is as follows:

Configure via modparams a set of devices per network.
EX: options networks=o2ib0(ib0,ib1,ib2)
  -> This represent the local IB bond.

Configure a set of targets in the form:
  -> <nid> <ip 1> <ip 2> ... <ipN>
  -> Where N < LNET_MAX_INTERFACES and  LNET_MAX_INTERFACES == 16;
  -> each NID <nid> in the target configuration is reachable via
     the configured set of interfaces.
EX:
10.1.1.2@o2ib0 10.1.1.2 10.1.1.3 10.1.1.4
When a node attempts to send a message to 10.1.1.2@o2ib0, it can
send the first message to 10.1.1.2 interface, then the next one
to 10.1.1.3 and the third to 10.1.1.4.  Then it round robins over
these interfaces.

This configuration gets passed to the LND and stored.
In the o2iblnd, it creates a set of what they term routes,
on connection establishment.  Each route is between a local
interface from the bond specified and a remote target IP.
The maximum number of routes is the greater of the number of
local interfaces or the number of configured target IP addresses.
The route list per connection is built in Round Robin.

When sending, an RPC messages is sent over one of these routes.

The next route in the list is chosen in RR for each outgoing RPC.

RPC replies are sent over the same connection.

This patch excludes the channel bonding interface used to
configure targets.  This will be added by a subsequent
patch which creates a DLC compliant interface.

Signed-off-by: NAGUCHI Yoshinori <y-naguchi@jp.fujitsu.com>
Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I371262261217dcf520f3e200bdfe83379410f592